Tweak auditor to handle aliased paths
Auditor local/remote/db maps are keyed on path, so aliased paths break the auditor. This change allows the DB object to be found even when the local file is an alias of the db/remote item, and changes the local map to key on unaliased path.

(this way, comparing local and remote map keys should not generate spurious move events)